Revamp fire interface to fully handle fire actions
This commit wraps the various ways of triggering a direction chooser in
yet another layer of api that handles cycling between and selection of
arbitrary actions. The challenge here is that there are many ways to
trigger a direction chooser, which may involve extremely idiosyncratic
setup that is not controlled internal to the direction chooser. So, to
switch from e.g. a firestorm targeter to an iceblast wand targer is
something that simply can't be done inside the direction chooser UI
without some kind of extremely major rewrite. Instead, what this does is
rebuild the custom targeter for each action while cycling in a way that
looks like a seamless UI to the player.

In the long run, there may be ways of gradually refactoring bits and
pieces of code to make this less painful. For example, if every action
were only called via the quiver::action api, then the intervening could
could be slimmed down quite a bit, with most messaging etc refactored
into actions. But, this is not practical to do all at once.  I have
however done this for throwing as a kind of prototype, since this was
basically the simplest case. (But, there's still more
refactoring/simplification to do even there.)

For untargeted spells, right now in the fire interface you basically
just get a variant of the "just looking" targeter. However, I think
there are better things to do for many of these spells that will come in
a future commit.