Improve handling of disabled actions when cycling fire
In some cases, depending on the details of the action, disabled actions
can't get to a direction chooser at all (case: sandblast with 0 stones),
and so cycling to them in the fire prompt would exit the prompt.  This
just skips them for UI consistency. You can still force quiver them, and
they may still exit the prompt when doing so (depends on the action).