Simplify launching?
This gives launchers their own quiver that stably retains the last
quivered ammo for that launcher, so that the main quiver slot can hold
something else.

tldr: tab(/v) always attacks with the main weapon if possible (/aims
with the main weapon for non-melee). Shift-tab(/f) always fires the
quivered action (/aims the quivered action). You still use the main
quiver to change the launcher ammo where relevant. (Which is pretty much
only for slings.)

Associated changes in detail:

* tab always fires a wielded launcher (as long as its quiver is not
  empty). After this change, tab is much closer to a "primary attack"
  button triggering whatever is wielded, and you have to work at it to
  get tab (as opposed to shift-tab) to fire the main quiver. (This can
  still happen if you wield a non-weapon, probably, or by messing with
  autofight settings.)
* launchers can be fired using v even if the main quiver has something
  else, similar to reaching weapons. This is kind of odd, but doing
  this as another case of CMD_EVOKE_WIELDED allowed me to reuse a
  bunch of code and also allowed me not to complicate the lua interface
  beyond what was already there for reaching.
* quivering relevant ammo in the main quiver sets it for the launcher.
  This currently the only way to change the launcher quiver. Quivering
  nothing also force-clears the launcher quiver.
* wielding a launcher autoswitches only the launcher quiver to the
  correct ammo. (Vestiges of old quiver code still track what was last
  used and select that, but this matters only for slings. More of the
  old quiver code has been removed.)
* if the quivers are different, the launcher ammo is displayed to the
  right of the wielded weapon name (potentially overwriting some of it).
  If they are the same, only the main quiver version is shown. (In this
  case, tab and shift-tab have identical results -- but shift-tab will
  only fire a launcher if that launcher's ammo is showing in the main
  quiver.)

This is, perhaps, a step to having a full "primary action" slot
(associated by default with whatever is wielded, but where anything
could be quivered) and a secondary action slot corresponding to the
current regular quiver.